DCSANet/requirements.txt
accelerate==1.0.1
albumentations==1.4.18
astunparse==1.6.3
fvcore==0.1.5.post20221221
hydra-core==1.3.2
iopath==0.1.10
numpy==1.24.4
omegaconf==2.3.0
onnx==1.21.0
onnxruntime==1.24.4
onnxsim==0.6.2
opencv_python_headless==4.12.0.88
Pillow==12.2.0
pycocotools==2.0.7
Pygments==2.20.0
safetensors==0.7.0
tabulate==0.9.0
termcolor==3.3.0
terminaltables==3.1.10
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.67.1
xformers==0.0.35
yapf==0.43.0
